+++ src/utils/event.js
@@ -1,0 +1,6 @@
+export const humanizeTime = (date) => {
+  const hours = date.getHours().toString().padStart(2, `0`);
+  const minutes = date.getMinutes().toString().padStart(2, `0`);
+
+  return `${hours}:${minutes}`;
+};
